Enable sandi meter
Taking care of the code readability,
we try to apply the Sandi Metz rules. [1]

To automate this process, we enable sandi meter gem. [2]
We've used it before but for some reason we had to disable it.
We also add configuration files:
- sandi_meter/.sandi_meter - contains ignored files
- sandi_meter/config.yml - contains percentages for each rule

We had to ignore the migrations folder and one seed file.
This should not have a negative effect on the application.

[1]: https://thoughtbot.com/blog/sandi-metz-rules-for-developers
[2]: https://github.com/makaroni4/sandi_meter