Update Jekyll icon
The embedded Jekyll icon was cropped
because it doesn't have a viewbox attribute,
and the width and height attributes are changed during embedding.
After updating the SVG with the viewBox attribute,
the embedded icon looks good.

In addition, the IDs in the SVG file were not unique,
which may cause problems with colors and gradients in embedding.
I took care of that too.