Bump bootstrap from version 3 to 5
As part of our transition to rails 7 and updating our
dependencies. We want to move from bootstrap 3 to bootstrap 5.

This change will allow us to move away from using the npm package
of bootstrap to using the gem and in the future we will be able
to switch from using sprockets to using importmaps.

In this commit we also change the way we import css from using
requires to using imports in scss. This change is requires so
that bootstrap will be able to use mixins.