Use embedded SVG for social media icons
We use embedded SVGs to list the technologies.
We want to do the same thing with the social media list.

To pass the tests,
I just had to make sure the SVG tag was visible in the page's source,
not a file the `.svg` extension.

This implementation is not perfect.
Instead of a helper method,
it would probably be better to create a class
that supports embedded SVG.
It's on our list of changes.