Correct map generation
Loading the javascript in a different way forces us to make
some changes to the way we load leaflet in our javascript.
We use import instead of having leaflet preloaded in the script.
While making this change we also remove the dependency of this
script on jQuery. In the future we would like to remove jQuery
completely from the project.