Add a new category
I added the linters and code formaters category
to our technology list.

I actually added 2 new enum values:
- linter
- code_formatter

But we want to pull out technologies
that fall into at least one of these two categories.
For this reason, I created scope.
Scope in the SQL query implementation has "or".
However, the collection we get contains technologies
from both categories (not in the same record),
which is why I used "and".