Inherit the default exclude list for Rubocop
Rubocop skips the vendor folder by default. [1]
However, our setup adds new paths to exclude.
In this way, we accidentally overwrite the default settings.
We don't want that to happen.
That's why we allow these settings to be merged. [2]

We didn't notice this problem before
because bundler locally installs gems
outside of the vendor folder (in gemset).

The problem surfaced while using the Github Actions workflow. [3]

[1]: https://github.com/rubocop/rubocop/blob/master/config/default.yml#L63:L67
[2]: https://docs.rubocop.org/rubocop/0.88/configuration.html#merging-arrays-using-inherit_mode
[3]: https://stackoverflow.com/questions/70818301/rubocop-allcops-exclude-not-working-in-github-actions#answer-70818366