Redesign of the technological part on the homepage
The technology part didn't look good.
Especially on mobile devices.

Because sometimes we have pictures on the right
and sometimes on the left,
on small screens (mobile phone)
the pictures began to appear one below the other
without a description.

To change this, I used the official example from Bootstrap page. [1]

In addition, I moved the right and left display way into partials.

[1]: https://getbootstrap.com/docs/5.3/examples/carousel/