Restructure Community index view
The Community index view has been restructured
to eliminate the need for conditional class tags
and improve the organization of the code.
Specifically, separate partials have been created for the left
and right versions of the community list items,
which previously relied on conditional tags.

This change simplifies the code and sets the stage
for further improvements to the Communities#index view.