Use Bootstrap accordion for technology list
I used Bootstrap's accordion component to display a categorized
list of technologies on the bottom of the page.
I set the accordion as collapsed to avoid overwhelming the user
with a long list of technologies.

This allows visitors to expand the accordion to explore
the specific technologies the individual knows.

This change improves the user experience by organizing
the information in a more clearer format.