Remove uglifier
I noticed that asset precompiling doesn't work properly
in production environment.
I found an issue about uglifier in the importmap-rails repo.
What DHH wrote makes sense to me. [1]

In the future, we may try to use terser. [2]
We will also need to add a miniracer [3]
or other supported JavaScript interpreter. [4]
For now, we've gotten rid of server-side JavaScript.
We've removed the hidden server-side JavaScript (for now).

[1]: https://github.com/rails/importmap-rails/issues/5
[2]: https://github.com/ahorek/terser-ruby
[3]: https://github.com/rubyjs/mini_racer
[4]: https://github.com/rails/execjs