Create a people names migration
We want to transfer the simple responsibility
of the full name to the database.
We want to store only first and last name in the database.
The full name is to be generated as a virtual column. [1]

We also don't want a fullname column, but full_name.

Based partly on the GitLab documentation, [2]
I also created unit tests for this migration.

[1]: https://api.rubyonrails.org/classes/ActiveRecord/ConnectionAdapters/Table.html
[2]: https://docs.gitlab.com/ee/development/testing_guide/testing_migrations_guide.html