Add default value from database to enum list
After recent changes to the default Rails configuration, [1]
we started getting the following message
when loading seeds into the database:

> PG::NotNullViolation: ERROR:
> null value in column "category" of relation "technologies"
> violates not-null constraint

Our code here has not changed.
Everything worked before.

When creating a record,
we did not set any value from the enum,
because in the database for this integer column,
the default value was 0.
When the record was created,
we set the appropriate enum value.

To solve the problem, we added the unknown value to the enum.
We don't want to make many changes.
It was the simplest solution.

[1]: https://github.com/fractalsoft/fractalsoft.org/commit/2f29170b26547df7b9cd0a3c6cc717aa7109a817