Fix importmaps for custom js
https://trello.com/c/O1mjkISk

After successfully deploying a test company application to `fly.io`,
we noticed that our custom js files weren't working.
The problem was how we import custom js files.
We use relative imports the custom js files in the `application.js`
which works in the development but it breaks in production. [1]

To solve this problem we need to pin all files
in custom fractalsoft directory in `importmaps.rb`
and change the import in `application.js`. [2][3]

[1]: https://stackoverflow.com/questions/70548841/how-to-add-custom-js-file-to-new-rails-7-project
[2]: https://github.com/rails/importmap-rails/issues/66
[3]: https://github.com/rails/importmap-rails/issues/67