Remove the duplication from importmap
In our importmap setup, we pinned jquery twice.
This didn't cause any error,
however it can be confusing as
we were using two different jquery versions.

I removed the older version of the pinned library.

For reasons unknown to me yet,
I wasn't able to use the source on "ga.jspm.io"
but had to use it from "cdn.jsdelivr.net".
Otherwise I was getting "$ is not defined".

BTW I updated the jquery version to the latest one.