Fix sending mail from contact form
https://trello.com/c/O1mjkISk

When we try to send an email from the contact form
we get several problems.
We used the `mandrill_mail` and `mailchimps`
but we didn't get the mail to the mailbox.
We decided to stop using with `mandrill_mail` and `mailchimps`.

For sending emails we wanted to use `gmail`
but we had problems with `username` and `password`:
> Gmail SMTP send - 535-5.7.8 Username and Password not accepted. [1][2]

Finally, we omitted the external services
and used our SMTP server directly and it works.
We can send an email and we get it to the our mailbox.🎉

[1]: https://stackoverflow.com/questions/23137012/535-5-7-8-username-and-password-not-accepted
[2]: https://support.google.com/domains/answer/9437157?hl=en