Replace imagemagick with vips
So far, we've used MiniMagick for image manipulation, [1]
which is a Ruby wrapper for ImageMagick.
The only time we process images is when we run `rails db:seed`.
For such small application, ImageMagick is quite heavy.
It turned out that CarrieWave, which we use to upload files, [2]
has alternative solutions. [3]
The ruby-vips is a Ruby gem which binding [4]
the libvips image processing library. [5]
This solution seems lighter, so we want to experiment with it. [6]

[1]: https://github.com/minimagick/minimagick
[2]: https://github.com/carrierwaveuploader/carrierwave
[3]: https://github.com/carrierwaveuploader/carrierwave/pull/2500
[4]: https://github.com/libvips/ruby-vips
[5]: https://www.libvips.org/
[6]: https://github.com/libvips/ruby-vips#benchmarks