Fix CoderDojo logo image processing
The CoderDojo logo is quite wide and not very tall.
Therefore, after changing the image processing library,
we had to make some concessions.
I changed the image processing a bit so that it wasn't so problematic.
Thanks to this, the logo is not cut off.
It's just a little different, but still correct.