Remove `webdrivers` gem
The dependabot wants to update the `webdrivers` gem to 5.3.1.
With this update, `selenium-webdriver` is automatically dropped
from 4.12.0 to 4.10.0 version.
This is because the 5.3.0 `webdrivers` gem
is restricted to versions 4.0 - 4.10 for `Selenium`. [1]

When we update `webdrivers` to the latest version,
we drop the `selenium-webdriver` version to 4.10.
This will cause the next step of dependabot to try
to update `selenium-webdriver` to the latest version,
which will lower the version of `webdrivers` and so on...

In the `webdrivers` commit message is information
to stop using `webdrivers`
when we have `Ruby` 3+ and `Selenium` 4.11+. [2]

> Selenium itself now manages drivers by default [3]

[1]: https://github.com/titusfortner/webdrivers/blob/main/CHANGELOG.md
[2]: https://github.com/titusfortner/webdrivers/commit/c7a6f3b97031891692999a9b5cc69e55e8da0f17
[3]: https://www.selenium.dev/documentation/selenium_manager