Fix spaces between sentences
https://trello.com/c/MV9Jyte6

We had bad text formatting for community descriptions.
Community descriptions are strings
with the sentences separated by newline characters `"\n"`.
Some longer sentences are consist of more than one line
with a newline character.

We have defined in html_minifier.rb to remove `"\r\n"`
and this causes joining of words before and after this character.
This results in missing spaces between sentences
or words where the newline character was.

Changing `"\r\n"` to space in html_minifier.rb solved the problem.