Add an effect for member pictures
Team member pictures are links to detailed pages
about the selected person.

But when we hover over the picture,
only the cursor changes its appearance.
Nothing else indicates that anything will happen
after we click on the picture.

The picture needs to be more interactive
to encourage users to click on it.