Update person decorator to use virtual columns
Modified the person decorator's methods to utilize
the newly added virtual columns for social media URLs.

Previously, we relied on Ruby code to generate the appropriate URL
based on person identifiers. With the introduction of virtual columns,
this implementation has been updated
to use the virtual columns directly.

This change streamlines the code and ensures that decorated links
to social media profiles now use the virtual columns,
simplifying the process of generating
and displaying social media links with associated SVG icons.