Fix name of field in seeds
Among other things,
we changed the names of fields in the database for the person model.
We changed blog to blog_url and website to website_url.

But we didn't change the names of the fields
for assign_attributes in seeds.
The problem occurs when we set up a new application environment.

I'm getting information about aborted loading of seeds
because of incompatibility name of fields during `rails db:seed`.