Update technology record for PostGIS
I added an SVG file with the brand logo. [1]
Because this version of the file is much better
than the one we generated, they replace the old logo.

The advantage of this is that the page should load faster
(embedded SVG) and the logo should not have artifacts on the sides.

[1]: https://postgis.net/