Install vips for Zeitwerk workflow on CI
Running Zeitwerk check on CI failed and we saw in the logs:

> Run bundle exec rails zeitwerk:check
> Hold on, I am eager loading the application.
> rails aborted!
>
> LoadError: Could not open library 'vips.so.42':
> vips.so.42: cannot open shared object file:
> No such file or directory. (LoadError)
>
> Could not open library 'libvips.so.42':
> libvips.so.42: cannot open shared object file:
> No such file or directory.
>
> Searched in <system library path>, /usr/lib, /usr/local/lib,
> /opt/local/lib
...
> Error: Process completed with exit code 1.

Since we have our dev environment prepared locally,
we were not aware of the problem until we launched the workflow on CI.

Installing the appropriate vips library should help.

In the future, we may think about a way to cache system packages. [1]

[1]: https://github.com/awalsh128/cache-apt-pkgs-action