Refactor usage of Redcarpet
We were missing tests to check
the conversion from Markdown to HTML.
While writing the missing tests,
I found out that some implementations for using Redcarpet didn't work.

We had added some Redcarpet extensions, [1]
but they did not load in the application when we used them.
To be more specific,
the Redcarpet extensions were not enabled in the application
where we were using them.

This was the reason for refactoring the code.

The tests that check for correct conversion of Markdown
to HTML with Redcarpet pass.

[1]: https://www.writesoftwarewell.com/how-to-render-markdown-views-in-rails/