Add Sort by Position
We have a method that returns contributions for a team member.

On the FractalSoft page
we want to have the team member contributions displayed
in a custom order.
This order can be imposed by the value of the post position.

We want the result of this method to be sorted
by the ascending value of the contribution positions.

Adding Order by position to the method,
gives us the desired result.