Add libyaml to Dockerfile
When upgrading Rails to 7.1.2,
the psych gem was automatically installed in version 5.1.1.1.
While deploying the application to fly.io we get the following error:

> Gem::Ext::BuildError: ERROR: Failed to build gem native extension.

The deployment failed.

The reason for this error
is that the psych gem requires the libyaml library.

Adding this library to the Dockerfile solved the problem.