Change configuration for Commonmarker
I need to change the configuration for Commonmarker [1]
to render the HTML code from Markdown correctly.
The current configuration causes that Markdown to ommit
part of the code with raw HTML,
among other things for Youtube movies.

I need to find the correct render
and extension for Commonmarker to be able to render raw HTML
and enable other features of Markdown.

The combination of `unsafe: true` render
and `tagfilter: false` extension
helps Markdown with HTML to return HTML code.
But this combination disables some other default extensions like:
tables, autolinks, task list and striked text.

Disabled extensions have to be enabled directly
in the Commonmarker configuration.

[1]: https://github.com/gjtorikian/commonmarker/blob/main/README.md