Fix HtmlMinifier to handle frozen strings
On the production server,
we received the following message in the logs:

```
FrozenError - can't modify frozen String: "":
  app/middleware/html_minifier.rb:37:in `block (2 levels) in compress'
  app/middleware/html_minifier.rb:36:in `block in compress'
  app/middleware/html_minifier.rb:35:in `compress'
  app/middleware/html_minifier.rb:26:in `call'
```

We didn't see this in development mode because we only enable
the `HtmlMinifier` middleware in production.
Additionally, the problem only occurred during redirects.

To fix this, I added unit tests for the middleware [1]
and reproduced the case where the response contains a frozen string.

The fix skips the chunk that is frozen.
And the tests will help us protect ourselves
from a similar error in the future.

[1]: https://stackoverflow.com/questions/17506567/testing-middleware-with-rspec