Refactor create method in ContactFormsController
https://trello.com/c/Z9QslfBi

The user can send an email from two places.
The first is at the bottom of the home page
in the Contact us section.
The second is on the contact page.

The current behavior of sending an email
is that the user doesn't know if his email has been sent.
After clicking the button to send an email,
he is simply redirected to the home page and that is all.
This behavior is the same for both places used to send an email.

We want to the user to see the message
about the result of sending an email
and to stay in at the same place where he sends an email.