Fix flaky tests
When running `bundle exec rspec spec --seed 1` the tests fail.

In my opinion,
the cause of the failed tests is that in some cases they depend on
the `I18n.locale` set in the tests.
Some tests were missing settings with the correct locale for them,
and depending on the order in which the tests are run,
they sometimes get the correct locale and sometimes not.

The `bundle exec rspec spec --seed 1` is the case
where the tests get wrong locale from other tests configuration.

Setting the correct locale for these tests directly
will fix the problem.