Do not load Leaflet library
We are not currently using Leaflet on our website. [1]
We got what we wanted to achieve in plain HTML using OpenStreetMap. [2]
If we want to redesign and beautify it in the future,
we will use Leaflet.

We can be inspired by examples from the official Leaflet.js website. [3]
We can use our logo to mark the office on the map
or use geolocation to calculate how far it is from our office.
We really have a lot of ideas.

[1]: https://leafletjs.com/
[2]: https://www.openstreetmap.org/
[3]: https://leafletjs.com/examples.html