Fix display of team members on their page
Each team member is introduced on their page.
The entire description of the person
is contained in a kind of frame.
This page includes a personal picture,
a short introduction, sometimes with a quote,
and a link to their Github and/or social media,
as well as their skills.

These elements are displayed side by side in full view.
Manipulating the width of the window causes
the arrangement of these elements to change.

And it changes, but not what we want.
In some window widths the person's introduction
covers a part of the picture.
And there are some window widths
that skills are behind the frame
that includes the whole description.

The correct display appears
for smaller widths starting around 553 px.

After making changes,
the person introduction will no longer overlap the image,
and the skills portion will remain within the frame.