Fix display of Contact Us section on the home page
At the bottom of the home page we have the Contact Us section.
On the full view it displays the contact form
and next to it the company addresses in line.

If we change the window width to smaller,
the arrangement remains the same,
but from some width value it is unreadable
and looks very strange for the user.

For the smallest bootstrap grid value it starts to look clean
and comfortable for the user again.

The introduced changes bring order
for displaying part of Contact Us section
and behave more stable while changing window width.