Fix display of contact form
We have a page with contact form to send a message from the user.
This page contains, among other things,
the company addresses, the map with the office locations
and the contact form.

When the window width changes,
the arrangement of these elements breaks
and looks unprofessional.

I added some bootstrap classes
to fix the arrangement the contact page elements.