Fix display of a team member card
On the main page of team members
we have introduction for each person.
For each person is displayed his image, name, short introduction,
button to the page with full information about this person
and link to Github or social media.
These data are organized inside bootstrap card.

Now we have defined a constant space
between the short introduction and the button.
This causes the button to have a different position
depending on the size of the introduction.

It is visible because on the page
the team members are displayed by two cards next to each other.
This makes it look messy.

I decided to remove our setting for the space
between the introduction and the button
and add a setting for the Bootstrap
to correct the arrangement inside the person card.