Add BusinessContactForm Model
https://trello.com/c/khqHa7MN

A business contact form is different
from an ordinary contact form.
It is more specific
and it is used for special kind of contact.
It is intended for contact
for the potential client who want to cooperate with us.
More details could help to start dialog about cooperation.
The client can present his problem, area of interest etc.
and we can have picture
of what kind of help he needs and working area.