Add BusinessContactFormController
https://trello.com/c/khqHa7MN

Depending on the result of sending a message,
it is necessary to redirect the user to the correct place
with information about this result.
If the user successfully sent a message,
he should know that everything went well.
If sending a message fails,
we need to inform the user about it
and give him an opportunity to contact us again.