Move long translations to the Rails partials
One of our translations had a long value, spanning multiple lines.
Additionally, the normalization of translations was not facilitating
our work. To better organize the code and make maintenance easier,
I decided to adhere to the Rails naming convention.
I moved the translations from the en.yml and pl.yml files to partials.
Now, the default file `_thank_you.md` is used for all locales,
including English,
while the file `_thank_you.pl.md` is used only for the Polish language.

I used markdown files because the content used does not require HTML.
However, our Markdown linter noticed that the headline was missing
from the file. This is not true because there are only partials
and the headline is in the show file.
To make the linter happy, I added exceptions. [1]

[1]: https://github.com/DavidAnson/markdownlint#configuration