Create a GitHub Action workflow for tests in Ruby
https://trello.com/c/OxjfG6Co

We run our tests locally and with every commit push.
But before adding new code to the application,
especially before deploying to production,
we want to be more sure that everything works correctly.

Adding the ability to run tests on CI
helps us ensure that the application still works.
In our case, we need to connect to PostgreSQL,
so we need to create a PostgreSQL service container for this [1].

[1]: https://docs.github.com/en/actions/using-containerized-services/creating-postgresql-service-containers