Fix Github action workflow for tests
I added a github action workflow for ruby tests,
but not everything is going right.
While running the ruby tests action
I get an error and this workflow fails.
The first problem was setting up the database:
> bin/rails aborted!
> ActiveRecord::AdapterNotSpecified:
> The `development` database is not configured
> for the `development` environment.
> (ActiveRecord::AdapterNotSpecified)
> Available database configurations are: test

The problem was that
I didn't define the correct environment in the workflow step,
so the action used development.
Adding `RAILS_ENV: test` while setting up the database
and running the test solves this problem.

The second problem was with the specific test:
spec/models/image_spec.rb.
I get an error:
> An error occurred while loading ./spec/models/image_spec.rb.
> - Did you mean?
> rspec ./spec/models/person_spec.rb
> rspec ./spec/models/project_spec.rb
> Failure/Error: include CarrierWave::Vips

> LoadError:
> Could not open library 'vips.so.42': vips.so.42:
> cannot open shared object file: No such file or directory.
> Could not open library 'libvips.so.42': libvips.so.42:
> cannot open shared object file: No such file or directory.
> Searched in <system library path>, /usr/lib,
> /usr/local/lib, /opt/local/lib

The problem was that
we were using the ImageUploader in the Image model.
The ImageUploader includes CarrierWare::Vips, [1]
which requires the libvips library. [2]

Adding the installation of this library to the workflow solves the problem.
I also added updating the system package list
to make sure everything works correctly.

[1]: https://github.com/eltiare/carrierwave-vips
[2]: https://makandracards.com/makandra/604272-carrierwave-processing-images-with-libvips