Change the loading way for static file
https://trello.com/c/SCR4N4bL

We want to be able to deploy the application from Github action.
While trying to deploy the application from Github action
I got an error:

> Error error [statics] are not yet supported
> when deploying to machines;
> remove the [statics] section from fly.toml

By moving the loading of static files to Rails,
we can remove the statics section from the fly.toml file.

Changing the way the static file is loaded solves this problem.