Add the intl-tel-input to importmap
https://trello.com/c/GEDbEMW1

In the business contact form
we have input for phone number
that the user can fill in.

For a better user experience
we want to use the intl-tel-input JavaScript plugin. [1]
This plugin has many settings
that provide better user experience
while entering the phone number.
Among other things,
this plugin allows the user to select their country based on
the country flag and country name.
With this selection, the dialing code is added,
which helps to enter the correct phone number.

Now all the user has to do is enter the phone number.

[1]: https://github.com/jackocnr/intl-tel-input