Add phone attributes for BusinessContactForm
https://trello.com/c/GEDbEMW1

Using the intl-tel-input JavaScript plugin,
we make it easier for the user to enter the phone number.

Now the user doesn't have to think
about how to enter the phone number.
The plugin helps to set the dialing code
and the user only has to enter the number.

We need to know the country and area code
to correctly interpret the user's number.
This information is important
to contact the potential customer correctly.
Besides the phone number
we need the attributes for country and area code.