Update intl-tel-input
To keep things up to date we want to update dependencies from our
importmaps. Intl tel input was out of date so we want to update it.
We run `bin/importmap update` which bumps the js version and puts the
current version of the package into vendor. We also update css to match
the version and remove the css hacks that were but are no longer necesary
with the newer version.